% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_by.R
\name{cumsum_by}
\alias{cumsum_by}
\alias{cumsum_by.data.frame}
\alias{cumsum_by.formula}
\title{Cumulative sums for database lists}
\usage{
cumsum_by(x, ...)

\method{cumsum_by}{data.frame}(x, ids, vars, by, suffix = "_cumul", diff = FALSE, ...)

\method{cumsum_by}{formula}(x, data, ...)
}
\arguments{
\item{x}{A data frame containing or a formula, depending on the method. See
the description above to design the formula.}

\item{...}{Further arguments passed among methods.}

\item{ids}{A character vector with the names of the variables used to
identify every experimental object (usually the combination of treatment
and repetition).}

\item{vars}{A character vector with the names of the variables used to
calculate the cumulative sum.}

\item{by}{A character value used as temporal variable.}

\item{suffix}{A character value added to the name of the variable in the
output data frame.}

\item{diff}{A logical value. If true, the lagged differences (function
\code{\link[=diff]{diff()}}) instead of the cumulative sums (function \code{\link[=cumsum]{cumsum()}}) will be
calculated.}

\item{data}{A data frame required in the formula method.}
}
\value{
The input data frame with the cumulative sums as additional variables.
Note that the sorting of rows may change in the process.
}
\description{
Observations needing the calculation of a cumulative sum, for instance in
germination experiments, will require to be done by treatment, combining
identifiers (e.g. treatment and repetition).

In the formula method you set in the left term the numeric variable which
has to be acummulated. Note that you can insert in the left term multiple
variables separating them by a plus symbol (\code{'+'}).
On the right term you set the sorting variable (usually time, Date or
integer) as the first term followed by identifiers.

This function is checking the uniqueness of the temporal variable per
treatment.
For ties the function will retrieve an error message.
}
