\name{eval_model}
\alias{eval_model}

\title{Evaluating Models according to Residuals}

\description{
Describe!
}
\usage{
eval_model(vars, match, group, pred, obs, FUN=list())
}
\arguments{
\item{vars}{Character vector with the variables to be evaluated.}
\item{match}{Variable used to match samples in observed values.}
\item{group}{Variable used to group samples.}
\item{pred}{Data frame containing predicted variables.}
\item{obs}{Data frame containing observed variables.}
\item{FUN}{List with functions used for the evaluation.}
}
\details{
Detail!
}
\value{
A data frame.
}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}.
}
\examples{
library(cropgrowth)

## Make an example!
}
