% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pile.R
\name{pile}
\alias{pile}
\alias{pile,data.frame-method}
\alias{pile,formula-method}
\title{Pile Partitions of Growth Curves}
\usage{
pile(x, ...)

\S4method{pile}{data.frame}(x, time, biomass, ...)

\S4method{pile}{formula}(x, data, ...)
}
\arguments{
\item{x}{A data frame or a formula.}

\item{...}{Further arguments passed among methods.}

\item{time, biomass}{Character vectors indicating the names of time and
biomass variables in the input data frame.}

\item{data}{A data frame including time and biomass variables.}
}
\value{
A S4 object of class \linkS4class{piled_curves}.
}
\description{
Generating objects of class  \linkS4class{piled_curves}.

In the case of crops, sorting of data frames should be done from left to
right according to the bottom-up position of organs in the plants.

In the method \code{'formula'}, the terms at the left side are the sorted
biomass variables, while at the right side are the time variables.
}
\examples{
## Subset to one treatment
maize_model <- subset(maize_model, treatment == "T3")

## Method for data frame
maize_growth <- pile(x=maize_model, time=c("current_date","development_stage","tsum"),
	biomass=c("wst","wglv","wgld","yield"))

## Method for formula
maize_growth <- pile(wst + wglv + wgld + yield ~ current_date + development_stage + tsum,
	data=maize_model)

}
